updaters/unbound: Ensure that destination files exist
An issue I'm having with the recent updates is that NixOS brings
Unbound down because of the configuration changes, then can't bring
it back up because it's missing a file created by a freshly-enabled
updater; meanwhile, the updater can't run because the DNS resolver
is down.  To break this cycle, create a blank file during activation
so that Unbound can boot up normally, let the updater run, and then
load in the properly updated file.